requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.17
pydantic==2.9.2
sqlmodel==0.0.22
SQLAlchemy==2.0.35
alembic==1.13.2
pillow==10.4.0
# Optional (install if you want OCR/barcodes early)
# pytesseract==0.3.13
# pyzbar==0.1.9
# opencv-python==4.10.0.84
# autogen (pin the version you prefer)
# pyautogen==0.2.34
